Update README.md
The main objectives of this re-write is to 
1) Make it easier for someone finding Nordpool Diff to understand what it does
2) Make it easier for someone who decides to try out Nordpool Diff to get started. This includes providing more guidance and recommendations, and more practical applications.
3) Make it easier for someone who wants to tweak Nordpool Diff but isn't great in maths to understand how to Think about different ways of configuring it.
4) Improving readability in general

I've toned down the emphasis for using ENTSO-E over Nordpool, since over the last year: 1) Nordpool is and official integration since HA 2024.12, meaning HACS is not needed while HACS. 2) I had to stop using ENTSO-E because it had a bug for a few months causing it to be unusable, and no one fixed it for quite a while, i.e. there are doubts to how well maintained the ENTSO-E integration is.